==> 02-prepX-may-aug.Rmd <==
---
title: "alltogether"
author: "Alexandros Papageorgiou"
date: "21 November 2018"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

options(scipen=999)
options(digits=3)


```


find mode

```{r}
find_mode <- function(myinput){
    names(sort(summary(as.factor(myinput)), decreasing=T)[1:1])
}

columns_to_mode <- c("channelGrouping", "country", "city", "subContinent", "deviceCategory", "operatingSystem", "browser")



df_mode_aug <- df_thinner_4 %>% filter(date < as_datetime("2018-08-16")) %>%    group_by(fullVisitorId) %>% summarise_at(.vars =columns_to_mode,  find_mode) 

df_mode_aug_action <-  df_thinner_4 %>% 
    filter(date < as_datetime("2018-08-16")) %>%    group_by(fullVisitorId) %>% 
    summarise(act_type_max=mean(act_type_max)) %>% mutate(act_type_avg=ifelse(is.na(act_type_max), 0, act_type_max)) %>% 
    select(-act_type_max) %>% 
    left_join(df_mode_aug) 


```


number of purchases by number of visits

```{r}
# buys_by_nbrvisits <- new_data_final %>% 
#     group_by(fullVisitorId) %>% 
#     summarise(number_sessions=n(), number_Trans=sum(hasRev)) %>% 
#     group_by(number_sessions) %>% mutate(numberids_with_)
#     mutate(nbr_of_trans_per_id=)
#     group_by(count) %>% summarise(count_total=n(), sumoftransactions=sum(sumTrans)) %>% 
#     arrange(desc(count_total)) 
# 
# rate <- buys_by_nbrvisits %>% mutate(rate=sumoftransactions/count_total)


```



df_til_aug

```{r}



df_til_aug <- df_thinner_4 %>% filter(date < as_datetime("2018-08-16")) %>% 
  group_by(fullVisitorId) %>%
  dplyr::summarise(
    totRev_aug = sum(transactionRevenue),
    transactions_aug = sum(transactionRevenue > 0),
    avg_transaction_aug=mean(transactionRevenue),
    sessions_aug = dplyr::n(),
    Ginternal_aug = max(isGoogleinternal > 0),
    dates_aug = list(date),
    #conv_dates = list(map(date, function(x) if (.$transactionRevenue >0) x )),
    max_session_aug = max(date),
    max_conversion_aug = max(ifelse(transactionRevenue > 0, date, as.Date("2018-01-01")))) %>% 
    mutate(max_conversion_aug = as_datetime(max_conversion_aug)) %>%
    mutate(dates_aug=map(dates_aug, as_datetime)) %>%  
    mutate(session_post_June =as_datetime(max_session_aug) > as_datetime("2018-06-15")) %>% 
    mutate(conv_post_June=max_conversion_aug > as_datetime("2018-06-15")) 

```

join trainV2_afterMay17 and testV2 for oneTrain


```{r}
#source("train-datav2-nohits.R")

#write_rds(trainv2_converters, "trainv2_converters.rds")
#trainv2_converters <- read_rds("trainv2_converters.rds")

# joint <- df_til_aug %>% left_join(trainv2_converters) %>% 
#     mutate(totalTrans_before=ifelse(is.na(totalTrans_before), 0, totalTrans_before)) %>% 
#     mutate(sumTrans_before=ifelse(is.na(sumTrans_before), 0, sumTrans_before)) 

trainv2_converters_fromMay17 <- read_csv("data-files/trainv2_converters_fromMay17.csv")

aug_12m_joint <- df_til_aug %>% left_join(trainv2_converters_fromMay17)

```




ids with transactions after aug (the target var)


```{r}

trans_aug_to_oct <- read_rds("data-files/hit_type_df_users_possibly.rds") %>% filter(date > as_datetime("2018-08-15")) %>% 
    mutate(transactionRevenue=as.numeric(transactionRevenue)) %>% 
    mutate(transactionRevenue=ifelse(is.na(transactionRevenue), 0, transactionRevenue)) %>% 
  group_by(fullVisitorId) %>%
  dplyr::summarise(
    totRev_test = sum(transactionRevenue),
    transactions_test = sum(transactionRevenue > 0),
    avg_transaction_test=mean(transactionRevenue),
    sessions_test = n())
    
    
  #   Ginternal = max(isGoogleinternal > 0),
  #   dates = list(date),
  #   #conv_dates = list(map(date, function(x) if (.$transactionRevenue >0) x )),
  #   max_session = max(date),
  #   max_conversion = max(ifelse(transactionRevenue > 0, date, as.Date("2018-01-01")))
  # ) %>%
  # mutate(max_conversion = as_datetime(max_conversion)) %>%
  #   mutate(dates=map(dates, as_datetime)) %>%  
  #   mutate(session_post_June =as_datetime(max_session) > as_datetime("2018-06-15")) %>% 
  #   mutate(conv_post_June=max_conversion > as_datetime("2018-06-15")) 


```

join 3

```{r}

x_and_y_aug <- aug_12m_joint %>% left_join(trans_aug_to_oct)

```

join mode

```{r}

x_and_y_mode_aug <- x_and_y_aug %>% left_join(df_mode_aug_action)

```

columns configure


```{r}



x_and_y_arranged_aug <- x_and_y_mode_aug %>%  select(fullVisitorId,
                                     act_type_avg:browser,
                                     sessions=sessions_aug,
                                     transactions=transactions_aug,
                                     Ginternal=Ginternal_aug,
                                     totalTrans_before,
                                     transactions_test,
                                     totRev_aug,
                                     avg_transaction_aug,
                                     dates_aug,
                                     max_session_aug, 
                                     max_conversion_aug, 
                                     sumTrans_before,
                                     totRev_test,
                                     avg_transaction_test,
                                     sessions_test) %>% 
    dplyr::mutate_at(vars(transactions_test, totRev_test, avg_transaction_test, sessions_test), replace_na, 0) %>% 
    #mutate(act_type_max=as.integer(act_type_max)) %>% 
    mutate(Ginternal=ifelse(Ginternal==0, F, T)) %>%
    mutate(transactions_test=ifelse(transactions_test==0, F, T)) %>% 
    mutate(city=ifelse(city=="NA's", NA, city)) %>% 
    mutate(totalTrans_before=ifelse(is.na(totalTrans_before), 0 , totalTrans_before))
    
vars_to_factor <- c("channelGrouping",
"country",
"city",
"subContinent",
"deviceCategory",
"operatingSystem",
"browser")



x_and_y_arranged_aug <-x_and_y_arranged_aug %>% mutate_at(vars(vars_to_factor), as.factor)
                        
```



```{r}
final_df <- x_and_y_arranged_aug %>% select(act_type_avg:transactions_test)  %>%  
    select(-city) %>% 
    mutate(transactions_test=ifelse(transactions_test==T, "converted", "deferred")) %>%
    mutate(transactions_test=as.factor(transactions_test)) 

write_rds(final_df, "data-files/final_df.rds")

#write_csv(final_df, "final_df.csv")

```


```{r}

# final_df2 <- x_and_y_arranged_aug %>% select(fullVisitorId:transactions_test)  %>%  
#     select(-city) %>% 
#     mutate(transactions_test=ifelse(transactions_test==T, "converted", "deferred")) %>%
#     mutate(transactions_test=as.factor(transactions_test))

```
